Move get_session_data into own file
