+++ webapp/app/forms/flows/eligibility_step_chooser.py
@@ -1,5 +1,6 @@
 from flask_babel import _
 
+from app.forms.session_data import get_session_data
 from app.forms.steps.eligibility_steps import EligibilityStartDisplaySteuerlotseStep, \
     IncomeOtherDecisionEligibilityInputFormSteuerlotseStep, \
     IncomeOtherEligibilityFailureDisplaySteuerlotseStep, \
@@ -86,7 +87,7 @@ class EligibilityStepChooser(StepChooser):
             If no such step is found, returns the first step in the flow.
         """
         current_step_idx = self.step_order.index(current_step_name)
-        stored_data = self._get_session_data()
+        stored_data = get_session_data(self.session_data_identifier, default_data=self.default_data())
         for possible_previous_step_idx in range(current_step_idx - 1, 0, -1):
             possible_previous_step = self.steps[self.step_order[possible_previous_step_idx]]
             if possible_previous_step.is_previous_step(current_step_name, stored_data):
