+++ webapp/app/forms/flows/step_chooser.py
@@ -5,6 +5,7 @@ from werkzeug.exceptions import abort
 
 from app import app
 from app.forms.flows.multistep_flow import deserialize_session_data
+from app.forms.session_data import get_session_data
 from app.forms.steps.steuerlotse_step import SteuerlotseStep, RedirectSteuerlotseStep
 
 
@@ -43,23 +44,10 @@ class StepChooser:
         else:
             return None
 
-    def _get_session_data(self, session_data_identifier=None, ttl: Optional[int] = None):
-        if session_data_identifier is None:
-            session_data_identifier = self.session_data_identifier
-        serialized_session = session.get(session_data_identifier, b"")
-
-        if self.default_data():
-            stored_data = self.default_data() | deserialize_session_data(serialized_session,
-                                                                         ttl)  # updates session_data only with non_existent values
-        else:
-            stored_data = deserialize_session_data(serialized_session, ttl)
-
-        return stored_data
-
     def get_correct_step(self, step_name) -> SteuerlotseStep:
         if self._get_possible_redirect(step_name):
             return RedirectSteuerlotseStep(self._get_possible_redirect(step_name), endpoint=self.endpoint)
-        stored_data = self._get_session_data()
+        stored_data = get_session_data(self.session_data_identifier, default_data=self.default_data())
 
         # By default set `prev_step` and `next_step` in order of definition
         return self.steps[step_name](
